+++ example/lib/main.dart
@@ -5,32 +5,129 @@ void main() {
   runApp(MyApp());
 }
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
+  @override
+  _MyAppState createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  late ThemeMode _themeMode;
+  late IconData iconToggle;
+
+  initState() {
+    super.initState();
+    _themeMode = ThemeMode.system;
+    iconToggle = Icons.toggle_on;
+  }
+
+  toggleThemeMode() {
+    if (_themeMode == ThemeMode.light) {
+      _themeMode = ThemeMode.dark;
+      iconToggle = Icons.toggle_off;
+    } else if (_themeMode == ThemeMode.dark) {
+      _themeMode = ThemeMode.light;
+      iconToggle = Icons.toggle_on;
+    } else {
+      if (Theme.of(context).brightness == Brightness.light) {
+        _themeMode = ThemeMode.dark;
+        iconToggle = Icons.toggle_off;
+      } else {
+        _themeMode = ThemeMode.light;
+        iconToggle = Icons.toggle_on;
+      }
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      theme: ThemeData(
-        backgroundColor: Colors.white,
-      ),
-      darkTheme: ThemeData(
-        backgroundColor: Colors.black,
-      ),
+      theme: ThemeData.light(),
+      darkTheme: ThemeData.dark(),
+      themeMode: _themeMode,
       home: Scaffold(
-        appBar: AppBar(title: Text('Dart Code Viewer Example')),
-        body: DartCodeViewer(r'''
+        appBar: AppBar(
+          title: Text('Dart Code Viewer Example'),
+          actions: [
+            IconButton(
+              icon: Icon(iconToggle),
+              onPressed: () {
+                setState(() {
+                  toggleThemeMode();
+                });
+              },
+            ),
+          ],
+        ),
+        body: DartCodeViewer(DartCode.template),
+      ),
+    );
+  }
+}
 
+class DartCode {
+  static const template = '''
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'dart_code_viewer_theme.dart';
-import 'prehighlighter.dart';
 
+import 'dart_code_viewer_theme.dart';
+import 'pre_highlighter.dart';
+
+/// A code viewer for the dart language.
+///
+/// A code viewer can be used to display dart code. By default the [DartCodeViewer]
+/// gives you a Theme based code view. If you are using a [ThemeMode] that is light
+/// than you will get the light option. Note that the default background of the code
+/// viewer is based off [ColorScheme.background].
+///
+/// Supplying a non-null [data] String is required as input.
+///
+/// Requires one of its ancestors to be a [Material] widget.
+///
+/// Requires one of its ancestors to be a [MediaQuery] widget. Typically, these
+/// are introduced by the [MaterialApp] or [WidgetsApp] widget at the top of
+/// your application widget tree.
+///
+/// {@tool dartpad --template=stateless_widget_scaffold}
+///
+/// ![A dart_code_viewer example for light mode.]
+/// (https://github.com/JoseAlba/dart_code_viewer/images/import_example)
+///
+/// Here is an example of a small string that shows up as Dart code in a flutter
+/// application.
+///
+/// ```dart
+/// @override
+/// Widget build(BuildContext context) {
+///   return DartCodeViewer(r'class DartCodeViewer extends StatelessWidget {}');
+/// }
+/// ```
+/// {@end-tool}
+///
+/// See also:
+///  * [DartCodeViewerTheme] and [DartCodeViewerThemeData] for information about
+///  controlling the visual appearance of the DartCodeViewer.
+///  * [Code viewer online tool](https://romannurik.github.io/SlidesCodeHighlighter/)
+///  is a useful tool that lets you choose the color for each different style.
+///  On the left side you put your example code and on the right you can choose
+///  the colors you want the code viewer to display.
+///    background => backgroundColor
+///    plain text => baseStyle
+///    Punctuation => punctuationStyle
+///    String, values => stringStyle
+///    Keywords, tags => keywordStyle
+///    Comments => commentStyle
+///    Types => classStyle
+///    Numbers => numberStyle
+///    Declarations => constantStyle
+///  * [MediaQuery], from which the default height and width factor is obtained.
 class DartCodeViewer extends StatelessWidget {
+  /// DartCodeViewer requires a [String] that will be the code shown within the
+  /// code viewer. This should be dart code and it is preferable if you use a raw
+  /// string by adding an r before the string.
   const DartCodeViewer(
     this.data, {
-    Key key,
+    Key? key,
     this.baseStyle,
     this.classStyle,
     this.commentStyle,
@@ -44,37 +141,41 @@ class DartCodeViewer extends StatelessWidget {
     this.showCopyButton,
     this.height,
     this.width,
-  }) : assert(data != null, 'A non-null String must be provided to a DartCodeViewer widget.'),
-      super(key: key);
+  }) : super(key: key);
 
+  /// Create a DartCodeViewer based of one [TextStyle]. Optional [Color] parameters
+  /// which change the TextStyle color for that highlighter type.
+  ///
+  /// The default [TextStyle] is [RobotoMono].
+  ///
+  /// Useful parameter when you want to use one [TextStyle].
   factory DartCodeViewer.textColor(
     String data, {
-    TextStyle textStyle,
-    Color baseColor,
-    Color classColor,
-    Color commentColor,
-    Color constantColor,
-    Color keywordColor,
-    Color numberColor,
-    Color punctuationColor,
-    Color stringColor,
-    Color backgroundColor,
-    Text copyButtonText,
-    bool showCopyButton,
-    double height,
-    double width,
+    TextStyle? textStyle,
+    Color? baseColor,
+    Color? classColor,
+    Color? commentColor,
+    Color? constantColor,
+    Color? keywordColor,
+    Color? numberColor,
+    Color? punctuationColor,
+    Color? stringColor,
+    Color? backgroundColor,
+    Text? copyButtonText,
+    bool? showCopyButton,
+    double? height,
+    double? width,
   }) {
-    TextStyle _defaultCodeStyle = textStyle ?? GoogleFonts.robotoMono(fontSize: 12);
     return DartCodeViewer(
       data,
-      baseStyle: _defaultCodeStyle.copyWith(color: baseColor),
-      classStyle: _defaultCodeStyle.copyWith(color: classColor),
-      commentStyle: _defaultCodeStyle.copyWith(color: commentColor),
-      constantStyle: _defaultCodeStyle.copyWith(color: constantColor),
-      keywordStyle: _defaultCodeStyle.copyWith(color: keywordColor),
-      numberStyle: _defaultCodeStyle.copyWith(color: numberColor),
-      punctuationStyle: _defaultCodeStyle.copyWith(color: punctuationColor),
-      stringStyle: _defaultCodeStyle.copyWith(color: stringColor),
+      baseStyle: textStyle?.copyWith(color: baseColor),
+      classStyle: textStyle?.copyWith(color: classColor),
+      commentStyle: textStyle?.copyWith(color: commentColor),
+      constantStyle: textStyle?.copyWith(color: constantColor),
+      keywordStyle: textStyle?.copyWith(color: keywordColor),
+      numberStyle: textStyle?.copyWith(color: numberColor),
+      punctuationStyle: textStyle?.copyWith(color: punctuationColor),
+      stringStyle: textStyle?.copyWith(color: stringColor),
       backgroundColor: backgroundColor,
       copyButtonText: copyButtonText,
       showCopyButton: showCopyButton,
@@ -83,13 +184,14 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
+  /// Common code viewer highlighter for [ThemeMode.light].
   factory DartCodeViewer.light(String data) {
     return DartCodeViewer.textColor(
       data,
-      baseColor: Colors.blueGrey.shade100,
+      baseColor: Colors.blueGrey.shade800,
       classColor: Colors.purple.shade500,
       commentColor: Colors.pink.shade600,
-      constantColor: Colors.indigo.shade50,
+      constantColor: Colors.indigo.shade500,
       keywordColor: Colors.indigo.shade500,
       numberColor: Colors.red.shade700,
       punctuationColor: Colors.blueGrey.shade800,
@@ -98,6 +200,7 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
+  /// Code viewer light alternative for [ThemeMode.light].
   factory DartCodeViewer.lightAlt(String data) {
     return DartCodeViewer.textColor(
       data,
@@ -113,6 +216,7 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
+  /// Common code viewer highlighter for [ThemeMode.dark].
   factory DartCodeViewer.dark(String data) {
     return DartCodeViewer.textColor(
       data,
@@ -128,6 +232,7 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
+  /// Code viewer dark alternative for [ThemeMode.dark].
   factory DartCodeViewer.darkAlt(String data) {
     return DartCodeViewer.textColor(
       data,
@@ -143,6 +248,7 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
+  /// Code viewer highlighter with a great dark design for [ThemeMode.dark].
   factory DartCodeViewer.designDark(String data) {
     return DartCodeViewer.textColor(
       data,
@@ -158,6 +264,7 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
+  /// Code viewer highlighter for Google IO 2017.
   factory DartCodeViewer.io17(String data) {
     return DartCodeViewer.textColor(
       data,
@@ -173,6 +280,7 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
+  /// Code viewer highlighter for Google IO 2019.
   factory DartCodeViewer.io19(String data) {
     return DartCodeViewer.textColor(
       data,
@@ -188,6 +296,7 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
+  /// Code viewer highlighter for Flutter Interact 2019.
   factory DartCodeViewer.flutterInteract19(String data) {
     return DartCodeViewer.textColor(
       data,
@@ -203,82 +312,136 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
-  final TextStyle baseStyle;
-  final TextStyle classStyle;
-  final TextStyle commentStyle;
-  final TextStyle constantStyle;
-  final TextStyle keywordStyle;
-  final TextStyle numberStyle;
-  final TextStyle punctuationStyle;
-  final TextStyle stringStyle;
+  /// The string that is transformed into code. This is a required variable.
   final String data;
 
-  final Color backgroundColor;
-  final Text copyButtonText;
-  final bool showCopyButton;
+  /// The text style for the plain text in code.
+  final TextStyle? baseStyle;
 
-  final double height;
-  final double width;
+  /// The text style for the code types in the code.
+  ///
+  /// For example:
+  /// * The class name.
+  /// * StatelessWidget and StatefulWidget.
+  final TextStyle? classStyle;
+
+  /// The text style for the commented out code.
+  final TextStyle? commentStyle;
+
+  /// The text style for the constant style code.
+  final TextStyle? constantStyle;
+
+  /// The text style for keywords. For example:
+  /// * else
+  /// * enum
+  /// * export
+  /// * external
+  /// * factory
+  /// * false
+  final TextStyle? keywordStyle;
+
+  /// The text style for numbers within the code.
+  final TextStyle? numberStyle;
+
+  /// The text style for punctuation code like periods and commas.
+  final TextStyle? punctuationStyle;
+
+  /// The text style for Strings. For example the data when using the [Text] widget.
+  final TextStyle? stringStyle;
+
+  /// The background Color of the code. By default it is [Theme.of(context).colorScheme.background].
+  final Color? backgroundColor;
+
+  /// The text shown in the copy button by default it is 'COPY ALL'.
+  final Text? copyButtonText;
+
+  /// Shows copy button that lets user copy all the code as a raw string. By
+  /// default the button is showing.
+  final bool? showCopyButton;
+
+  /// The height of the [DartCodeViewer] by default it uses the [MediaQuery.of(context).size.height]
+  final double? height;
+
+  /// The width of the [DartCodeViewer] by default it uses the [MediaQuery.of(context).size.width]
+  final double? width;
 
   @override
   Widget build(BuildContext context) {
-    final codeTextStyle = GoogleFonts.robotoMono(
-        fontSize: 12 * MediaQuery.of(context).textScaleFactor,
+    final codeTextStyle = Theme.of(context).textTheme.bodyText1;
+
+    final lightModeOn = Theme.of(context).brightness == Brightness.light;
+
+    // These are defaults for the different types of text styles. The default
+    // returns two different types of styles depending on the brightness of the
+    // application.
+    final _defaultBaseStyle = codeTextStyle?.copyWith(
+      color: lightModeOn ? Colors.blueGrey.shade800 : Colors.blueGrey.shade50,
+    );
+    final _defaultClassStyle = codeTextStyle?.copyWith(
+      color: lightModeOn ? Colors.purple.shade500 : Colors.purple.shade200,
+    );
+    final _defaultCommentStyle = codeTextStyle?.copyWith(
+      color: lightModeOn ? Colors.pink.shade600 : Colors.pink.shade300,
+    );
+    final _defaultConstantStyle = codeTextStyle?.copyWith(
+      color: lightModeOn ? Colors.indigo.shade500 : Colors.yellow.shade700,
+    );
+    final _defaultKeywordStyle = codeTextStyle?.copyWith(
+      color: lightModeOn ? Colors.indigo.shade500 : Colors.cyan.shade300,
+    );
+    final _defaultNumberStyle = codeTextStyle?.copyWith(
+      color: lightModeOn ? Colors.red.shade700 : Colors.yellow.shade700,
+    );
+    final _defaultPunctuationalStyle = codeTextStyle?.copyWith(
+      color: lightModeOn ? Colors.blueGrey.shade800 : Colors.blueGrey.shade50,
+    );
+    final _defaultStringStyle = codeTextStyle?.copyWith(
+      color: lightModeOn ? Colors.green.shade700 : Colors.lightGreen.shade400,
     );
 
-    /// Uses device brightness to choose what type of theme to return.
-    final brightness = WidgetsBinding.instance.window.platformBrightness;
-    bool lightModeOn = brightness == Brightness.light;
+    final _defaultCopyButtonText = Text('COPY ALL');
+    final _defaultShowCopyButton = true;
 
-    TextStyle _defaultBaseStyle = codeTextStyle.copyWith(
-      color: lightModeOn ? Colors.blueGrey.shade100 : Colors.blueGrey.shade50
-    );
-    TextStyle _defaultClassStyle = codeTextStyle.copyWith(
-      color: lightModeOn ? Colors.purple.shade500 : Colors.purple.shade200
-    );
-    TextStyle _defaultCommentStyle = codeTextStyle.copyWith(
-      color: lightModeOn ? Colors.pink.shade600 : Colors.pink.shade300
-    );
-    TextStyle _defaultConstantStyle = codeTextStyle.copyWith(
-      color: lightModeOn ? Colors.indigo.shade50 : Colors.yellow.shade700
-    );
-    TextStyle _defaultKeywordStyle = codeTextStyle.copyWith(
-      color: lightModeOn ? Colors.indigo.shade500 : Colors.cyan.shade300
-    );
-    TextStyle _defaultNumberStyle = codeTextStyle.copyWith(
-      color: lightModeOn ? Colors.red.shade700 : Colors.yellow.shade700
-    );
-    TextStyle _defaultPunctuationalStyle = codeTextStyle.copyWith(
-      color: lightModeOn ? Colors.blueGrey.shade800 : Colors.blueGrey.shade50
-    );
-    TextStyle _defaultStringStyle = codeTextStyle.copyWith(
-      color: lightModeOn ? Colors.green.shade700 : Colors.lightGreen.shade400
-    );
-
-    Text _defaultCopyButtonText = Text(''
-      'COPY ALL',
-      style: Theme.of(context).textTheme.button.copyWith(
-        color: Colors.white,
-        fontWeight: FontWeight.w500,
-      ),
-    );
-    bool _defaultShowCopyButton = true;
-
-    DartCodeViewerThemeData dartCodeViewerThemeData = DartCodeViewerTheme.of(context);
+    var dartCodeViewerThemeData = DartCodeViewerTheme.of(context);
     dartCodeViewerThemeData = dartCodeViewerThemeData.copyWith(
-      baseStyle: baseStyle ?? dartCodeViewerThemeData.baseStyle ?? _defaultBaseStyle,
-      classStyle: classStyle ?? dartCodeViewerThemeData.classStyle ?? _defaultClassStyle,
-      commentStyle: commentStyle ?? dartCodeViewerThemeData.commentStyle ?? _defaultCommentStyle,
-      constantStyle: constantStyle ?? dartCodeViewerThemeData.constantStyle ?? _defaultConstantStyle,
-      keywordStyle: keywordStyle ?? dartCodeViewerThemeData.keywordStyle ?? _defaultKeywordStyle,
-      numberStyle: numberStyle ?? dartCodeViewerThemeData.numberStyle ?? _defaultNumberStyle,
-      punctuationStyle: punctuationStyle ?? dartCodeViewerThemeData.punctuationStyle ?? _defaultPunctuationalStyle,
-      stringStyle: stringStyle ?? dartCodeViewerThemeData.stringStyle ?? _defaultStringStyle,
-      backgroundColor: backgroundColor ?? dartCodeViewerThemeData.backgroundColor ?? Theme.of(context).colorScheme.background,
-      copyButtonText: copyButtonText ?? dartCodeViewerThemeData.copyButtonText ?? _defaultCopyButtonText,
-      showCopyButton: showCopyButton ?? dartCodeViewerThemeData.showCopyButton ?? _defaultShowCopyButton,
-      height: height ?? dartCodeViewerThemeData.height ?? MediaQuery.of(context).size.height,
-      width: width ?? dartCodeViewerThemeData.width ?? MediaQuery.of(context).size.width,
+      baseStyle:
+          baseStyle ?? dartCodeViewerThemeData.baseStyle ?? _defaultBaseStyle,
+      classStyle: classStyle ??
+          dartCodeViewerThemeData.classStyle ??
+          _defaultClassStyle,
+      commentStyle: commentStyle ??
+          dartCodeViewerThemeData.commentStyle ??
+          _defaultCommentStyle,
+      constantStyle: constantStyle ??
+          dartCodeViewerThemeData.constantStyle ??
+          _defaultConstantStyle,
+      keywordStyle: keywordStyle ??
+          dartCodeViewerThemeData.keywordStyle ??
+          _defaultKeywordStyle,
+      numberStyle: numberStyle ??
+          dartCodeViewerThemeData.numberStyle ??
+          _defaultNumberStyle,
+      punctuationStyle: punctuationStyle ??
+          dartCodeViewerThemeData.punctuationStyle ??
+          _defaultPunctuationalStyle,
+      stringStyle: stringStyle ??
+          dartCodeViewerThemeData.stringStyle ??
+          _defaultStringStyle,
+      backgroundColor: backgroundColor ??
+          dartCodeViewerThemeData.backgroundColor ??
+          Theme.of(context).colorScheme.background,
+      copyButtonText: copyButtonText ??
+          dartCodeViewerThemeData.copyButtonText ??
+          _defaultCopyButtonText,
+      showCopyButton: showCopyButton ??
+          dartCodeViewerThemeData.showCopyButton ??
+          _defaultShowCopyButton,
+      height: height ??
+          dartCodeViewerThemeData.height ??
+          MediaQuery.of(context).size.height,
+      width: width ??
+          dartCodeViewerThemeData.width ??
+          MediaQuery.of(context).size.width,
     );
 
     return DartCodeViewerTheme(
@@ -295,21 +458,28 @@ class DartCodeViewer extends StatelessWidget {
     );
   }
 
-  InlineSpan codifyString(String content, DartCodeViewerThemeData dartCodeViewerThemeData) {
-    List<TextSpan> textSpans = [];
-    final codeSpans = DartSyntaxPrehighlighter().format(content);
-    // Converting CodeSpan to TextSpan by first converting to String and then TextSpan.
+  InlineSpan codifyString(
+    String content,
+    DartCodeViewerThemeData dartCodeViewerThemeData,
+  ) {
+    var textSpans = <TextSpan>[];
+    final codeSpans = DartSyntaxPreHighlighter().format(content);
+    // Converting CodeSpan to TextSpan by first converting to a string and then TextSpan.
     for (final span in codeSpans) {
       textSpans.add(stringToTextSpan(span.toString(), dartCodeViewerThemeData));
     }
     return TextSpan(children: textSpans);
   }
 
-  TextSpan stringToTextSpan(String string, DartCodeViewerThemeData dartCodeViewerThemeData) {
+  TextSpan stringToTextSpan(
+    String string,
+    DartCodeViewerThemeData dartCodeViewerThemeData,
+  ) {
     return TextSpan(
       style: () {
-        String styleString = RegExp(r'codeStyle.\w*').firstMatch(string).group(0);
-        DartCodeViewerThemeData dartCodeViewerTheme = dartCodeViewerThemeData;
+        final String? styleString =
+            RegExp(r'codeStyle.\w*').firstMatch(string)?.group(0);
+        var dartCodeViewerTheme = dartCodeViewerThemeData;
 
         switch (styleString) {
           case 'codeStyle.baseStyle':
@@ -333,11 +503,20 @@ class DartCodeViewer extends StatelessWidget {
         }
       }(),
       text: () {
-        String textString = RegExp('\'.*\'').firstMatch(string).group(0);
-        String subString = textString.substring(1, textString.length - 1);
+        final textString = RegExp('\'.*\'').firstMatch(string)?.group(0);
+        final subString = textString!.substring(1, textString.length - 1);
         return decodeString(subString);
       }(),
     );
+  }
+
+  /// Read raw string as regular String. Converts Unicode characters to actual
+  /// numbers.
+  String decodeString(String string) {
+    return string
+        .replaceAll(r'\u000a', '\n')
+        .replaceAll(r'\u0027', '\'')
+        .replaceAll(r'\u0009', '\t');
   }
 }
 
@@ -351,7 +530,7 @@ class _DartCodeViewerPage extends StatelessWidget {
     final _plainTextCode = _richTextCode.toPlainText();
 
     void _showSnackBarOnCopySuccess(dynamic result) {
-      Scaffold.of(context).showSnackBar(
+      ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
           content: Text('Copied to Clipboard'),
         ),
@@ -359,9 +538,9 @@ class _DartCodeViewerPage extends StatelessWidget {
     }
 
     void _showSnackBarOnCopyFailure(Object exception) {
-      Scaffold.of(context).showSnackBar(
+      ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: Text('Failure to copy to clipboard: $exception'),
+          content: Text('Failure to copy to clipboard: \$exception'),
         ),
       );
     }
@@ -369,19 +548,12 @@ class _DartCodeViewerPage extends StatelessWidget {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        if (DartCodeViewerTheme.of(context).showCopyButton)
-          FlatButton(
-            color: Colors.white.withOpacity(0.15),
-            materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-            padding: const EdgeInsets.symmetric(horizontal: 8),
-            shape: const RoundedRectangleBorder(
-              borderRadius: BorderRadius.all(Radius.circular(4)),
-            ),
+        if (DartCodeViewerTheme.of(context).showCopyButton!)
+          ElevatedButton(
             onPressed: () async {
-              await Clipboard.setData(
-                  ClipboardData(text: _plainTextCode)
-              ).then(_showSnackBarOnCopySuccess)
-               .catchError(_showSnackBarOnCopyFailure);
+              await Clipboard.setData(ClipboardData(text: _plainTextCode))
+                  .then(_showSnackBarOnCopySuccess)
+                  .catchError(_showSnackBarOnCopyFailure);
             },
             child: DartCodeViewerTheme.of(context).copyButtonText,
           ),
@@ -397,10 +569,5 @@ class _DartCodeViewerPage extends StatelessWidget {
     );
   }
 }
-
-      ''',
-       ),
-      ),
-    );
-  }
+''';
 }