+++ lib/src/pre_highlighter.dart
@@ -1,15 +1,11 @@
-// Copyright 2019 The Flutter team. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
 import 'package:string_scanner/string_scanner.dart';
 
-abstract class SyntaxPrehighlighter {
+abstract class SyntaxPreHighlighter {
   List<CodeSpan> format(String src);
 }
 
-class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
-  DartSyntaxPrehighlighter() {
+class DartSyntaxPreHighlighter extends SyntaxPreHighlighter {
+  DartSyntaxPreHighlighter() {
     _spans = <_HighlightSpan>[];
   }
 
@@ -76,10 +72,10 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
     'bool',
   ];
 
-  String _src;
-  StringScanner _scanner;
+  late String _src;
+  late StringScanner _scanner;
 
-  List<_HighlightSpan> _spans;
+  late List<_HighlightSpan> _spans;
 
   @override
   List<CodeSpan> format(String src) {
@@ -93,19 +89,29 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
 
       for (final span in _spans) {
         if (currentPosition != span.start) {
-          formattedText
-              .add(CodeSpan(text: _src.substring(currentPosition, span.start)));
+          formattedText.add(
+            CodeSpan(
+              text: _src.substring(currentPosition, span.start),
+            ),
+          );
         }
 
-        formattedText
-            .add(CodeSpan(type: span.type, text: span.textForSpan(_src)));
+        formattedText.add(
+          CodeSpan(
+            type: span.type,
+            text: span.textForSpan(_src),
+          ),
+        );
 
         currentPosition = span.end;
       }
 
       if (currentPosition != _src.length) {
-        formattedText
-            .add(CodeSpan(text: _src.substring(currentPosition, _src.length)));
+        formattedText.add(
+          CodeSpan(
+            text: _src.substring(currentPosition, _src.length),
+          ),
+        );
       }
 
       return formattedText;
@@ -126,20 +132,20 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r'/\*(.|\n)*\*/'))) {
         _spans.add(_HighlightSpan(
           _HighlightType.comment,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
 
       // Line comments
       if (_scanner.scan('//')) {
-        final startComment = _scanner.lastMatch.start;
+        final startComment = _scanner.lastMatch!.start;
 
         var eof = false;
         int endComment;
         if (_scanner.scan(RegExp(r'.*\n'))) {
-          endComment = _scanner.lastMatch.end - 1;
+          endComment = _scanner.lastMatch!.end - 1;
         } else {
           eof = true;
           endComment = _src.length;
@@ -162,8 +168,8 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r'r".*"'))) {
         _spans.add(_HighlightSpan(
           _HighlightType.string,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
@@ -172,8 +178,8 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r"r'.*'"))) {
         _spans.add(_HighlightSpan(
           _HighlightType.string,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
@@ -182,8 +188,8 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r'"""(?:[^"\\]|\\(.|\n))*"""'))) {
         _spans.add(_HighlightSpan(
           _HighlightType.string,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
@@ -192,8 +198,8 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r"'''(?:[^'\\]|\\(.|\n))*'''"))) {
         _spans.add(_HighlightSpan(
           _HighlightType.string,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
@@ -202,8 +208,8 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r'"(?:[^"\\]|\\.)*"'))) {
         _spans.add(_HighlightSpan(
           _HighlightType.string,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
@@ -212,8 +218,8 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r"'(?:[^'\\]|\\.)*'"))) {
         _spans.add(_HighlightSpan(
           _HighlightType.string,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
@@ -222,16 +228,21 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r'\d+\.\d+'))) {
         _spans.add(_HighlightSpan(
           _HighlightType.number,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
 
       // Integer
       if (_scanner.scan(RegExp(r'\d+'))) {
-        _spans.add(_HighlightSpan(_HighlightType.number,
-            _scanner.lastMatch.start, _scanner.lastMatch.end));
+        _spans.add(
+          _HighlightSpan(
+            _HighlightType.number,
+            _scanner.lastMatch!.start,
+            _scanner.lastMatch!.end,
+          ),
+        );
         continue;
       }
 
@@ -239,8 +250,8 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r'[\[\]{}().!=<>&\|\?\+\-\*/%\^~;:,]'))) {
         _spans.add(_HighlightSpan(
           _HighlightType.punctuation,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
@@ -249,18 +260,18 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
       if (_scanner.scan(RegExp(r'@\w+'))) {
         _spans.add(_HighlightSpan(
           _HighlightType.keyword,
-          _scanner.lastMatch.start,
-          _scanner.lastMatch.end,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
         ));
         continue;
       }
 
       // Words
       if (_scanner.scan(RegExp(r'\w+'))) {
-        _HighlightType type;
+        late _HighlightType type;
 
-        var word = _scanner.lastMatch[0];
-        if (word.startsWith('_')) {
+        var word = _scanner.lastMatch![0];
+        if (word!.startsWith('_')) {
           word = word.substring(1);
         }
 
@@ -274,15 +285,15 @@ class DartSyntaxPrehighlighter extends SyntaxPrehighlighter {
             word.startsWith('k') &&
             _firstLetterIsUpperCase(word.substring(1))) {
           type = _HighlightType.constant;
+        } else {
+          type = _HighlightType.base;
         }
 
-        if (type != null) {
-          _spans.add(_HighlightSpan(
-            type,
-            _scanner.lastMatch.start,
-            _scanner.lastMatch.end,
-          ));
-        }
+        _spans.add(_HighlightSpan(
+          type,
+          _scanner.lastMatch!.start,
+          _scanner.lastMatch!.end,
+        ));
       }
 
       // Check if this loop did anything
@@ -343,7 +354,7 @@ class _HighlightSpan {
 }
 
 class CodeSpan {
-  CodeSpan({this.type = _HighlightType.base, this.text});
+  CodeSpan({this.type = _HighlightType.base, required this.text});
 
   final _HighlightType type;
   final String text;
@@ -376,7 +387,6 @@ String _styleNameOf(_HighlightType type) {
     case _HighlightType.base:
       return 'baseStyle';
   }
-  return '';
 }
 
 String escape(String text) {